HSEARCH-3605 Move bridge mapping implementation classes from package "pojo.bridge.impl" to "pojo.bridge.mapping.impl"
