+++ mapper/pojo/src/main/java/org/hibernate/search/mapper/pojo/mapping/building/impl/PojoIdentityMappingCollectorImpl.java
@@ -15,11 +15,11 @@ import org.hibernate.search.mapper.pojo.bridge.IdentifierBridge;
 import org.hibernate.search.mapper.pojo.bridge.RoutingKeyBridge;
 import org.hibernate.search.mapper.pojo.bridge.mapping.programmatic.BridgeBuilder;
 import org.hibernate.search.mapper.pojo.logging.impl.Log;
-import org.hibernate.search.mapper.pojo.bridge.impl.IdentifierMapping;
-import org.hibernate.search.mapper.pojo.bridge.impl.PropertyIdentifierMapping;
-import org.hibernate.search.mapper.pojo.bridge.impl.ProvidedStringIdentifierMapping;
-import org.hibernate.search.mapper.pojo.bridge.impl.RoutingKeyBridgeRoutingKeyProvider;
-import org.hibernate.search.mapper.pojo.bridge.impl.RoutingKeyProvider;
+import org.hibernate.search.mapper.pojo.bridge.mapping.impl.IdentifierMapping;
+import org.hibernate.search.mapper.pojo.bridge.mapping.impl.PropertyIdentifierMapping;
+import org.hibernate.search.mapper.pojo.bridge.mapping.impl.ProvidedStringIdentifierMapping;
+import org.hibernate.search.mapper.pojo.bridge.mapping.impl.RoutingKeyBridgeRoutingKeyProvider;
+import org.hibernate.search.mapper.pojo.bridge.mapping.impl.RoutingKeyProvider;
 import org.hibernate.search.mapper.pojo.model.additionalmetadata.impl.PojoTypeAdditionalMetadata;
 import org.hibernate.search.mapper.pojo.model.path.impl.BoundPojoModelPath;
 import org.hibernate.search.mapper.pojo.model.path.impl.BoundPojoModelPathPropertyNode;
