+++ mapper/pojo/src/main/java/org/hibernate/search/mapper/pojo/mapping/impl/PojoIndexedTypeManager.java
@@ -18,8 +18,8 @@ import org.hibernate.search.engine.mapper.mapping.context.spi.MappingContextImpl
 import org.hibernate.search.engine.mapper.mapping.spi.MappedIndexManager;
 import org.hibernate.search.engine.mapper.scope.spi.MappedIndexScopeBuilder;
 import org.hibernate.search.engine.mapper.session.context.spi.DetachedSessionContextImplementor;
-import org.hibernate.search.mapper.pojo.bridge.impl.IdentifierMapping;
-import org.hibernate.search.mapper.pojo.bridge.impl.RoutingKeyProvider;
+import org.hibernate.search.mapper.pojo.bridge.mapping.impl.IdentifierMapping;
+import org.hibernate.search.mapper.pojo.bridge.mapping.impl.RoutingKeyProvider;
 import org.hibernate.search.mapper.pojo.dirtiness.impl.PojoImplicitReindexingResolver;
 import org.hibernate.search.mapper.pojo.dirtiness.impl.PojoReindexingCollector;
 import org.hibernate.search.mapper.pojo.model.spi.PojoCaster;
